jwt cookies integrated

+++ src/main/java/com/daas/resource/ProjectResource.java
@@ -5,6 +5,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import javax.ws.rs.Consumes;
+import javax.ws.rs.CookieParam;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
@@ -12,6 +13,7 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.core.Cookie;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
@@ -19,6 +21,7 @@ import com.amazonaws.auth.BasicAWSCredentials;
 import com.daas.aws.common.AmazonEC2Common;
 import com.daas.aws.common.AmazonIAMCommon;
 import com.daas.common.ConfFactory;
+import com.daas.common.DaaSConstants;
 import com.daas.model.Project;
 import com.daas.model.User;
 import com.daas.service.ProjectService;
@@ -26,6 +29,7 @@ import com.daas.service.UserService;
 import com.daas.service.impl.ProjectServiceImpl;
 import com.daas.service.impl.UserServiceImpl;
 import com.daas.util.DaasUtil;
+import com.daas.util.JWTUtil;
 import com.google.common.hash.HashFunction;
 import com.google.common.hash.Hashing;
 
@@ -47,14 +51,23 @@ public class ProjectResource {
 	@POST
 	@Path("/add/{user_id}")
 	@Consumes(MediaType.APPLICATION_JSON)
-	public Response addProject(Project project, @PathParam("user_id") String user_id) throws Exception{
+	public Response addProject(@CookieParam("daas-token") Cookie cookie, Project project, @PathParam("user_id") long user_id) throws Exception{
 
-		if(user_id==null || user_id.isEmpty() || user_id == "")
-			return Response.status(Response.Status.BAD_REQUEST).entity("Invalid user id").build();
+		if (cookie == null) {
+			return Response.serverError().entity("ERROR").build();
+		}
+
+		// validate jwt
+		String token = cookie.getValue();
+		boolean validToken = JWTUtil.parseJWT(token);
+
+		if(!validToken)
+			return Response.status(Response.Status.UNAUTHORIZED).entity("Unauthorized").build();
 
 		// check if valid user
-		if(!userService.userExists(Long.valueOf(user_id)))
+		if(!userService.userExists(user_id))
 			return Response.status(Response.Status.BAD_REQUEST).entity("Invalid User").build();
+
 
 		// check for mandatory fields, if null values
 		Map<String,Object> map = new  HashMap<String,Object>();
@@ -70,7 +83,7 @@ public class ProjectResource {
 			return Response.status(Response.Status.BAD_REQUEST).entity("Invalid IAM role name").build();
 
 		// generate project ID
-		project.setProject_id(hf.newHasher().putLong(System.currentTimeMillis()).putLong(Long.valueOf(user_id)).hash().toString());
+		project.setProject_id(hf.newHasher().putLong(System.currentTimeMillis()).putLong(user_id).hash().toString());
 
 
 		// check if this is User's first project
@@ -78,10 +91,9 @@ public class ProjectResource {
 		// if not. start another kubernetes container
 
 		// first project
-		User user = userService.read(Long.valueOf(user_id));
-
+		User user = userService.read(user_id);
 		String key = null;
-		if(user.getManagementEC2InstanceId() == null) {
+		if(user.getManagementEC2InstanceId() == DaaSConstants.TEMP_MGMT_EC2_INSTANCE_ID) {
 			// create keypair
 			AmazonEC2Common ec2 = new AmazonEC2Common(new BasicAWSCredentials(project.getCloud_access_key(), project.getCloud_secret_key()));
 			key = ec2.createEC2KeyPair(keypairName);
@@ -94,7 +106,7 @@ public class ProjectResource {
 
 		}
 
-		project.setUser_id(userService.read(Long.valueOf(user_id)));
+		project.setUser_id(user);
 		project.setDateCreated(System.currentTimeMillis());
 		project = projectService.create(project);
 
@@ -112,7 +124,18 @@ public class ProjectResource {
 	@GET
 	@Path("/{id}")
 	@Produces(MediaType.APPLICATION_JSON)
-	public Response getProject(@PathParam("id") String id){
+	public Response getProject(@CookieParam("daas-token") Cookie cookie, @PathParam("id") String id){
+
+		if (cookie == null) {
+			return Response.serverError().entity("ERROR").build();
+		}
+
+		// validate jwt
+		String token = cookie.getValue();
+		boolean validToken = JWTUtil.parseJWT(token);
+
+		if(!validToken)
+			return Response.status(Response.Status.UNAUTHORIZED).entity("Unauthorized").build();
 
 		Project project = projectService.read(id);
 
@@ -126,7 +149,18 @@ public class ProjectResource {
 	@PUT
 	@Path("/update")
 	@Consumes(MediaType.APPLICATION_JSON)
-	public Response updateProject(Project project){
+	public Response updateProject(@CookieParam("daas-token") Cookie cookie, Project project){
+
+		if (cookie == null) {
+			return Response.serverError().entity("ERROR").build();
+		}
+
+		// validate jwt
+		String token = cookie.getValue();
+		boolean validToken = JWTUtil.parseJWT(token);
+		
+		if(!validToken)
+			return Response.status(Response.Status.UNAUTHORIZED).entity("Unauthorized").build();
 
 		project = projectService.update(project);
 
@@ -140,7 +174,18 @@ public class ProjectResource {
 	@DELETE
 	@Path("/delete")
 	@Consumes(MediaType.APPLICATION_JSON)
-	public Response deleteProject(Project project){
+	public Response deleteProject(@CookieParam("daas-token") Cookie cookie, Project project){
+
+		if (cookie == null) {
+			return Response.serverError().entity("ERROR").build();
+		}
+
+		// validate jwt
+		String token = cookie.getValue();
+		boolean validToken = JWTUtil.parseJWT(token);
+
+		if(!validToken)
+			return Response.status(Response.Status.UNAUTHORIZED).entity("Unauthorized").build();
 
 		project = projectService.delete(project);
 
@@ -151,17 +196,17 @@ public class ProjectResource {
 	}
 
 
-	public static Project createFirstProject(Project project, String userId) throws IOException{
+	public static Project createFirstProject(Project project, Long userId) throws IOException{
 
 		// share Kube common AMI with User's AWS account
 		AmazonEC2Common ec2 = new AmazonEC2Common(new BasicAWSCredentials(ConfFactory.getPrivateConf().getString("vivek.aws.accessId"), ConfFactory.getPrivateConf().getString("vivek.aws.secretKey")));
 		ec2.shareAMIAcrossAccounts(amiId, project.getCloud_account_id());		
 
 		// create EC2 Kube Management server on User's AWS account
-		String instanceId = ec2.createEC2Instance(amiId, instanceType, project.getIam_admin_role(), securityGroupName, keypairName, project.getProject_id(), userId);
+		String instanceId = ec2.createEC2Instance(amiId, instanceType, project.getIam_admin_role(), securityGroupName, keypairName, project.getProject_id(), String.valueOf(userId));
 
 		// Setting the Instance ID for User
-		User user = userService.read(Long.parseLong(userId));
+		User user = userService.read(userId);
 		user.setManagementEC2InstanceId(instanceId);
 		userService.update(user);
